clang compiler and c99 support
