+++ examples/ackermann
@@ -12,9 +12,9 @@ def ackermann(m, n){
 write "acherman(0, 0) = "; write ackermann(0, 0); write "\n";
 write "acherman(1, 1) = "; write ackermann(1, 1); write "\n";
 write "acherman(2, 2) = "; write ackermann(2, 2); write "\n";
-write "acherman(3, 3) = "; write ackermann(3, 3); write "\n";
-write "acherman(3, 5) = "; write ackermann(3, 5); write "\n";
 write "acherman(3, 6) = "; write ackermann(3, 6); write "\n";
+write "acherman(3, 7) = "; write ackermann(3, 7); write "\n";
+write "acherman(3, 8) = "; write ackermann(3, 8); write "\n";
 
 % This test makes the FCallStack overflow - don't try, trust me.
-% write "acherman(3, 8) = "; write ackermann(3, 8); write "\n"; 
+% write "acherman(4, 1) = "; write ackermann(4, 1); write "\n";
